Refactor risk multiplier formatting into a functional component
